requirements.txt
fastapi==0.135.1
uvicorn==0.34.0
pandas==2.3.3
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
python-dotenv==1.2.1
azure-devops==7.1.0b4
pyotp==2.9.0
boto3==1.42.67